+++ CHANGELOG.md
@@ -20,3 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0-ea.1] - 2025-06-06
 - Remove dynamic UA.
+
+## [Unreleased]
+- Update `HttpRequestUtil.request` documentation